refactor: API 및 도메인 개선
변경 사항:
- ApiHeaders에서 사용자 로그인 ID를 X-USER-ID로 변경
- PointRepository에서 사용자 로그인 ID 관련 메서드 제거
- OrderPreparation 도메인 추가
- ProductLikeRepository에서 메서드 추가 및 수정
- ProductRepository에서 메서드 추가 및 수정
- ClockConfig 추가

비즈니스 로직:
- 사용자 ID 기반으로 포인트 및 좋아요 기능 개선

테스트:
- 관련 테스트 케이스 추가 및 수정

+++ apps/commerce-api/src/main/java/com/loopers/domain/product/Product.java
@@ -101,7 +101,7 @@ public class Product extends BaseEntity {
     return id != null && id.equals(getId());
   }
 
-  public void validateStockForOrder(Integer requestedQuantity) {
+  public void validateStockForOrder(Long requestedQuantity) {
     if (isNotAvailable()) {
       throw new CoreException(ErrorType.INSUFFICIENT_STOCK,
           String.format("상품 [%s]의 재고가 부족합니다.", name));
